[expo-updates][android] Isolate startup procedure code
