+++ packages/expo-updates/android/src/main/java/expo/modules/updates/UpdatesModule.kt
@@ -41,7 +41,7 @@ class UpdatesModule : Module() {
 
       val constants = mutableMapOf<String, Any>()
       try {
-        val constantsForModule = UpdatesController.instance.getConstantsForModule(context)
+        val constantsForModule = UpdatesController.instance.getConstantsForModule()
         val launchedUpdate = constantsForModule.launchedUpdate
         val embeddedUpdate = constantsForModule.embeddedUpdate
         val isEmbeddedLaunch = launchedUpdate?.id?.equals(embeddedUpdate?.id) ?: false
@@ -88,7 +88,6 @@ class UpdatesModule : Module() {
 
     AsyncFunction("reload") { promise: Promise ->
       UpdatesController.instance.relaunchReactApplicationForModule(
-        context,
         object : IUpdatesController.ModuleCallback<Unit> {
           override fun onSuccess(result: Unit) {
             promise.resolve(null)
@@ -116,7 +115,6 @@ class UpdatesModule : Module() {
 
     AsyncFunction("checkForUpdateAsync") { promise: Promise ->
       UpdatesController.instance.checkForUpdate(
-        context,
         object : IUpdatesController.ModuleCallback<IUpdatesController.CheckForUpdateResult> {
           override fun onSuccess(result: IUpdatesController.CheckForUpdateResult) {
             when (result) {
@@ -165,7 +163,6 @@ class UpdatesModule : Module() {
 
     AsyncFunction("fetchUpdateAsync") { promise: Promise ->
       UpdatesController.instance.fetchUpdate(
-        context,
         object : IUpdatesController.ModuleCallback<IUpdatesController.FetchUpdateResult> {
           override fun onSuccess(result: IUpdatesController.FetchUpdateResult) {
             when (result) {
